Refactor: decouple the logic for storing github history from the projectCrontributors.js
